refactor: move missing to typing module; underscore it
