+++ falcon/request.py
@@ -41,13 +41,13 @@ from falcon import request_helpers as helpers
 from falcon import util
 from falcon.constants import DEFAULT_MEDIA_TYPE
 from falcon.constants import MEDIA_JSON
-from falcon.constants import MISSING
-from falcon.constants import MissingOr
 from falcon.forwarded import _parse_forwarded_header
 from falcon.forwarded import Forwarded
 from falcon.media import Handlers
 from falcon.media.json import _DEFAULT_JSON_HANDLER
 from falcon.stream import BoundedStream
+from falcon.typing import MISSING
+from falcon.typing import MissingOr
 from falcon.typing import ReadableIO
 from falcon.typing import StoreArgument
 from falcon.util import deprecated
